+++ .github/workflows/ci-checks.yml
@@ -25,7 +25,7 @@ jobs:
     uses: Lightning-AI/utilities/.github/workflows/check-package.yml@v0.7.1
     with:
       actions-ref: v0.7.1
-      import-name: "pl_fairscale"
+      import-name: "lightning_fairscale"
       artifact-name: dist-packages-${{ github.sha }}
       testing-matrix: |
         {
