+++ .github/workflows/ci-testing.yml
@@ -59,7 +59,7 @@ jobs:
         pip list
 
     - name: Testing
-      run: coverage run --source pl_fairscale -m pytest src tests -v
+      run: coverage run --source lightning_fairscale -m pytest src tests -v
 
     - name: Statistics
       run: |
