+++ Makefile
@@ -10,7 +10,7 @@ test: clean
 	pip install -q -r tests/requirements.txt
 
 	# use this to run tests
-	python -m coverage run --source pl_fairscale -m pytest src tests -v --flake8
+	python -m coverage run --source lightning_fairscale -m pytest src tests -v --flake8
 	python -m coverage report
 
 docs: clean
