+++ docs/source/conf.py
@@ -29,7 +29,7 @@ sys.path.insert(0, os.path.abspath(_PATH_ROOT))
 SPHINX_MOCK_REQUIREMENTS = int(os.environ.get("SPHINX_MOCK_REQUIREMENTS", True))
 
 # alternative https://stackoverflow.com/a/67692/4521646
-spec = spec_from_file_location("pl_fairscale/__about__.py", os.path.join(_PATH_SOURCE, "pl_fairscale", "__about__.py"))
+spec = spec_from_file_location("lightning_fairscale/__about__.py", os.path.join(_PATH_SOURCE, "lightning_fairscale", "__about__.py"))
 about = module_from_spec(spec)
 spec.loader.exec_module(about)
