+++ src/lightning_fairscale/__init__.py
@@ -2,7 +2,7 @@
 
 import os
 
-from pl_fairscale.__about__ import *  # noqa: F401, F403
+from lightning_fairscale.__about__ import *  # noqa: F401, F403
 
 _PACKAGE_ROOT = os.path.dirname(__file__)
 _PROJECT_ROOT = os.path.dirname(_PACKAGE_ROOT)
