+++ src/lightning_fairscale/precision/__init__.py
@@ -1,0 +1,4 @@
+"""Precisions plugins."""
+
+from lightning_fairscale.precision.fully_sharded_amp import FullyShardedMixedPrecisionPlugin  # noqa: F401
+from lightning_fairscale.precision.sharded_amp import ShardedMixedPrecisionPlugin  # noqa: F401
