--- src/pl_fairscale/precision/__init__.py
@@ -1,4 +1,0 @@
-"""Precisions plugins."""
-
-from pl_fairscale.precision.fully_sharded_amp import FullyShardedMixedPrecisionPlugin  # noqa: F401
-from pl_fairscale.precision.sharded_amp import ShardedMixedPrecisionPlugin  # noqa: F401
