+++ tests/test_sharded_strategy.py
@@ -27,7 +27,7 @@ from pytorch_lightning.plugins import MixedPrecisionPlugin
 from pytorch_lightning.trainer.states import TrainerFn
 from torch import Tensor
 
-from pl_fairscale.strategies import DDPShardedStrategy, DDPSpawnShardedStrategy
+from lightning_fairscale.strategies import DDPShardedStrategy, DDPSpawnShardedStrategy
 
 
 class ModelWithAdamOptimizer(BoringModel):
@@ -232,7 +232,7 @@ def test_configure_ddp(tmpdir):
     trainer.predict(model, dataloaders=model.predict_dataloader())
 
 
-@mock.patch("pl_fairscale.strategies.DDPShardedStrategy._wrap_optimizers", autospec=True)
+@mock.patch("lightning_fairscale.strategies.DDPShardedStrategy._wrap_optimizers", autospec=True)
 @pytest.mark.parametrize("cls", [DDPShardedStrategy, DDPSpawnShardedStrategy])
 def test_custom_kwargs_sharded(_, cls):
     """Tests to ensure that if custom kwargs are passed, they are set correctly."""
@@ -242,14 +242,14 @@ def test_custom_kwargs_sharded(_, cls):
     strategy.parallel_devices = [Mock()]
     class_name = "sharded" if isinstance(strategy, DDPShardedStrategy) else "sharded_spawn"
 
-    with mock.patch(f"pl_fairscale.strategies.{class_name}.ShardedDataParallel", autospec=True) as mock_sharded:
+    with mock.patch(f"lightning_fairscale.strategies.{class_name}.ShardedDataParallel", autospec=True) as mock_sharded:
         strategy.configure_ddp()
     args, kwargs = mock_sharded.call_args
     assert "reduce_fp16" in kwargs
     assert kwargs["reduce_fp16"]
 
 
-@mock.patch("pl_fairscale.strategies.DDPShardedStrategy._wrap_optimizers", autospec=True)
+@mock.patch("lightning_fairscale.strategies.DDPShardedStrategy._wrap_optimizers", autospec=True)
 @pytest.mark.parametrize(("params", "expected_buffer_size"), [({}, 0), ({"reduce_buffer_size": 128}, 128)])
 @pytest.mark.parametrize("num_nodes", [1, 2])
 def test_custom_kwargs_sharded_reduce_buffer_size(_, params, expected_buffer_size, num_nodes):
@@ -260,7 +260,7 @@ def test_custom_kwargs_sharded_reduce_buffer_size(_, params, expected_buffer_siz
     strategy._lightning_module.trainer = Mock()
     strategy.parallel_devices = [Mock()]
 
-    with mock.patch("pl_fairscale.strategies.sharded.ShardedDataParallel", autospec=True) as mock_sharded:
+    with mock.patch("lightning_fairscale.strategies.sharded.ShardedDataParallel", autospec=True) as mock_sharded:
         strategy.configure_ddp()
     args, kwargs = mock_sharded.call_args
     assert "reduce_buffer_size" in kwargs
